ASSIGNMNT 9
Assignment 9: 
 
How will you create graphs using illustrator? Explain how one can edit labels or add symbols to the graphs. 
 
 
You can easily create graphs in Illustrator by selecting the Graph Tool (from the Tools panel), clicking on your Artboard and dragging to form the area size. In addition, you can type in the width and height of the graph, which is useful if you want to create graphs with specific dimensions in mind. To create a graph, first select one of the nine available graph tools. You will be able to change the type of graph a posteriori if desired. Then draw on the work plane the area you want the graph to occupy. You can also simply click on the artboard and specify the desired width and height settings. Enter the necessary data in the Graph Data window. The order of the data will vary depending on the type of graph chosen. Choose a cell and enter the desired value in the text capture field. You can also copy and paste data from a spreadsheet such as Excel into the Graph Data window. In addition, the first icon in the Graph data window allows you to import data from a spreadsheet file, If you reverse the values of the graph by mistake, the second icon allows you to reverse them. To apply the data to the 
graph, tap the icon depicting a checkmark. To display the data of an existing graph, choose in the OBJECT/ GRAPH / DATA menu. The Illustrator graph editor allows you to represent any type on your graph in any way you like, so every bar chart, mind map, or organizational chart looks exactly how you like it. Adding symbols or images to graphs You can customize the value bars with images. To do this, it is then a matter of selecting the desired image and creating a new symbol by choosing in the OBJECT/ GRAPH / SYMBOL menu. Click on Create then rename the symbol. To apply the symbol to a value bar, select the graph, right click and choose Bar. In the Bar window, select the desired symbol. The drop-down menu will allow you to choose whether you want the symbol to stretch, repeat, etc.